Fix startServer into a package instead of main
