+++ pkg/commands/commands.go
@@ -1,8 +1,7 @@
 package commands
 
 import (
-	"fmt"
-
+	"github.com/IBM/openkommander/pkg/rest"
 	"github.com/spf13/cobra"
 )
 
@@ -43,7 +42,7 @@ func RegisterCommands(rootCmd *cobra.Command) {
 		Use:   "server",
 		Short: "Start the OpenKommander REST API server",
 		Run: func(cmd *cobra.Command, args []string) {
-			startRESTServer()
+			rest.StartRESTServer()
 		},
 	}
 
@@ -59,33 +58,6 @@ func RegisterCommands(rootCmd *cobra.Command) {
 		topicsCmd,
 		serverCmd,
 	)
-}
-
-func restServerCommands() *cobra.Command {
-	serverCmd := &cobra.Command{
-		Use:   "server",
-		Short: "REST server commands",
-	}
-
-	startServerCmd := &cobra.Command{
-		Use:   "start",
-		Short: "Start the REST server",
-		Run: func(cmd *cobra.Command, args []string) {
-			fmt.Printf("Starting REST server...\n")
-		},
-	}
-
-	stopServerCmd := &cobra.Command{
-		Use:   "stop",
-		Short: "Stop the REST server",
-		Run: func(cmd *cobra.Command, args []string) {
-			fmt.Printf("Stopping REST server...\n")
-		},
-	}
-
-	serverCmd.AddCommand(startServerCmd, stopServerCmd)
-
-	return serverCmd
 }
 
 func topicsCommands() *cobra.Command {
